+++ CHANGELOG.md
@@ -6,6 +6,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 ## [Unreleased](https://github.com/cutechess/cutechess/tree/master)
 
 ### Added
+- dialog to show captured pieces (#813)
 
 ### Fixed